server_py/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets>=12.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
aiosqlite==0.19.0
pydantic==2.4.2
pydantic-settings==2.0.3
httpx==0.25.2